requirements.txt
absl-py==0.11.0
anykeystore==0.2
apex==0.1
bert-score==0.3.7
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cryptacular==1.5.5
cycler==0.10.0
cymem==2.0.5
datasets==1.1.3
defusedxml==0.6.0
dill==0.3.3
en-core-web-sm==2.3.1
filelock==3.0.12
fire==0.3.1
fsspec==0.8.5
future==0.18.2
gitdb==4.0.5
gitdb2==4.0.2
GitPython==3.1.11
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
hupper==1.10.2
idna==2.10
importlib-metadata==3.3.0
joblib==1.0.0
kiwisolver==1.3.1
langdetect==1.0.8
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
multiprocess==0.70.11.1
murmurhash==1.0.5
nlp-rake==0.0.2
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
packaging==20.8
pandas==1.1.5
PasteDeploy==2.1.1
pbkdf2==1.3
Pillow==8.0.1
pip==20.3.3
plac==1.1.3
plaster==1.0
plaster-pastedeploy==0.7
portalocker==2.0.0
preshed==3.0.5
protobuf==3.14.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.10.1
pyparsing==2.4.7
pyramid==1.10.5
pyramid-mailer==0.15.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytorch-lightning==1.1.0
pytz==2020.5
PyYAML==5.3.1
regex==2020.11.13
repoze.sendmail==4.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.6
sacrebleu==1.4.14
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.5.4
setuptools==51.0.0
six==1.15.0
sklearn==0.0
smmap==3.0.4
spacy==2.3.5
SQLAlchemy==1.3.22
srsly==1.0.5
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorboardX==1.8
termcolor==1.1.0
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.4.0+cu100
torchvision==0.5.0+cu100
tqdm==4.49.0
transaction==3.0.1
transformers==4.0.0
translationstring==1.4
typing-extensions==3.7.4.3
urllib3==1.26.2
velruse==1.1.1
venusian==3.0.0
wasabi==0.8.0
WebOb==1.8.6
Werkzeug==1.0.1
wheel==0.36.2
WTForms==2.3.3
wtforms-recaptcha==0.3.2
xxhash==2.0.0
zipp==3.4.0
zope.deprecation==4.4.0
zope.interface==5.2.0
zope.sqlalchemy==1.3